refactor: abstract class for irrationals instead of interface
