+++ README.md
@@ -74,7 +74,7 @@ Tests can be run using an IDE, or with the following command:
 ```./gradlew test```
 
 ### Resources
-When writing tests for irrationals, it can be useful to use a high-precision calculator, such as [this one](https://www.mathsisfun.com/scientific-calculator.html), to calculate expected values.
+When writing tests for irrationals, it can be useful to use a high-precision calculator, such as [this](https://www.mathsisfun.com/scientific-calculator.html) or [this](https://www.mathsisfun.com/calculator-precision.html), to calculate expected values.
 
 ## Linting
 Linting is done with [ktlint](https://ktlint.github.io/), using [this](https://github.com/jlleitschuh/ktlint-gradle) plugin.
